tests/wpt/web-platform-tests/dom/nodes/Document-contentType/contentType/contenttype_mimeheader_01.html

<!DOCTYPE html>
<title>Custom document.contentType === 'text/xml' when explicitly set to this value</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe style="display:none"></iframe>
<script>
async_test(function() {
  var iframe = document.getElementsByTagName('iframe')[0];
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "text/xml");
  }), false);
  iframe.src = "../support/contenttype_setter.py?type=text&subtype=xml";
});
</script>
