w3c-reports/12092014/mac-wk2/runned-tests/dom/nodes/Document-contentType/contentType/contenttype_mimeheader_01-expected.txt


FAIL Custom document.contentType === 'text/xml' when explicitly set to this value assert_equals: expected (string) "text/xml" but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-contentType/contentType/contenttype_mimeheader_01.html:11:18
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:843:37)
